Feat: 총지출(TotalExpense), 총수입(TotalIncome) 컴포넌트 구현.
