--- src/mine/useCount.js
@@ -1,8 +1,0 @@
-export default function useCount() {
-  function useMonthlyAmount(arr = []) {
-    const total = arr.reduce((s, v) => Number((s += v.amount)), 0)
-    return total.toLocaleString()
-  }
-
-  return { useMonthlyAmount }
-}
